walkways/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.11.2
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
